update: 更新文件 indicator.c  去除多余的类型转换：tlsVersion 是 uint32_t 类型，传递给 msgCb 时不再需要显式地转换为 int32_t，假设 msgCb 函数能够接受 uint32_t 类型。

+++ tls/feature/indicator/src/indicator.c
@@ -33,6 +33,6 @@ void INDICATOR_MessageIndicate(int32_t writePoint, uint32_t tlsVersion, int32_t 
         return;
     }
 
-    ctx->config.tlsConfig.msgCb(writePoint, (int32_t)tlsVersion, contentType, msg, msgLen, ctx, arg);
+    ctx->config.tlsConfig.msgCb(writePoint, tlsVersion, contentType, msg, msgLen, ctx, arg);
 }
 #endif /* HITLS_TLS_FEATURE_INDICATOR */
